various small tweaks: prefer unsigned over byte/int  (see ChangeLog)
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@7257 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/sound/SCC.hh
@@ -42,11 +42,11 @@ private:
 	virtual bool generateInput(int* buffer, unsigned num);
 
 	inline int adjust(signed char wav, byte vol);
-	byte readWave(byte channel, byte address, const EmuTime& time) const;
-	void writeWave(byte channel, byte offset, byte value);
+	byte readWave(unsigned channel, unsigned address, const EmuTime& time) const;
+	void writeWave(unsigned channel, unsigned offset, byte value);
 	void setDeformReg(byte value, const EmuTime& time);
-	void setFreqVol(byte address, byte value);
-	byte getFreqVol(byte address) const;
+	void setFreqVol(unsigned address, byte value);
+	byte getFreqVol(unsigned address) const;
 
 	static const int CLOCK_FREQ = 3579545;
